Make data_iterator_simple simpler
There was a documentation bug in data_iterator_simple. File cache was enabled by default but described as disabled. Enabling cache is always not a good idea as it forces to a tight coupling in examples in a same unit. In addition, the most dangerous thing is that users may not be aware of that their data augmentation in a data source function is applied **only once** although the important point of data augmentation is that we can provide more data even infinitely from a single data with random data perturbations. It becomes not random.